FormDataReference validator accepts invalid sandbox extensions https://bugs.webkit.org/show_bug.cgi?id=314948 rdar://176912134
Reviewed by Alex Christensen.

Check that sandbox extensions are not empty in FormDataReference::sandboxExtensionsAreSufficient.
We are checking that all extensions are successfully consumed in the FormDataReference constructor.
This patch is also removing the validator, since all validation is performed in the constructor.

* Source/WebKit/Platform/IPC/FormDataReference.serialization.in:
* Source/WebKit/Platform/IPC/FormDataReference.h:
(IPC::FormDataReference::sandboxExtensionsAreSufficient):
* Source/WebKit/Platform/IPC/FormDataReference.cpp: Added.
(IPC::FormDataReference::FormDataReference):
(IPC::FormDataReference::sandboxExtensionHandles const):
* Source/WebKit/Platform/IPC/FormDataReference.h:
(IPC::FormDataReference::sandboxExtensionsAreSufficient): Deleted.
(IPC::FormDataReference::sandboxExtensionHandles const): Deleted.
* Source/WebKit/Sources.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Originally-landed-as: 305413.939@safari-7624-branch (2427c3fef353). rdar://180428726
Canonical link: https://commits.webkit.org/316378@main
